i33 provide csv output (#48)
In addition to the existing PDF report, we now generate a CSV report. This is handy for importing a monthly report to a tracking spreadsheet. For me, it's usually to give someone a heads up that they're streaming a lot while not on WIFI etc. They pay the cost, but it's a nice courtesy to provide :)

The first header value in a new table is **Surrounded by Asterisks** for readability
The layout of the file mirrors the PDF report (some small tweaks were made the the PDF)
We generate both PDF and CSV every time.

Also...
* example bill.toml usage no longer adds up to Bill & Ted's favorite number
* patch up README for csv functionality